CDK-139: Adding location to DatasetDescriptors.
This adds a location URL to the DatasetDescriptor, which is accessible
through getLocation(). This also adds location(String) and location(URI)
and hasLocation() methods to the Builder and changes the schema(String)
to be a URI rather than a schema literal. To set a schema literal from a
String, schemaLiteral(String) was added.