CDK-139: Adding Configuration to DatasetDescriptor.
The last commit's addition of an optional location to the
DatasetDescriptor enables the MetadataProvider to determine the location
of Datasets and pass that location to a DatasetRepository. But, the
repository implementation was still dependent on having the correct
FileSystem. Adding the Configuration to the DatasetDescriptor as well
allows the FileSystemDatasetRepository to construct the correct
FileSystem for a descriptor, from the location URI and the
Configuration.