CDK-139: Adding list method to MetadataProvider.
If the DatasetRepository no longer determines the locations of data
sets, and instead relies on getting the location from the
MetadataProvider, then the repository can't list the available data
sets. The MetadataProvider determines what data sets are available, so
it should have the list() method. This is also mroe consistent with the
Hive implementation, where the repository already delegated to the
MetadataProvider.

The DatasetRepository interface still has exists and list methods, but
these should delegate to the MetadataProvider unless they need to be
hooked. For instance, if a Dataset has metadata but has not yet been
written to.