CDK-139: Separating common tests out of FS.
This extracts tests that should pass on all DatasetRepository
implementations into the TestDatasetRepositories set. The FS tests now
inherit from this add add other tests, like making sure that the data
directory exists after a dataset is created. Minor updates to the FS
repository are included where tests found bugs.

A major update to the FS repository and common tests is the new
MemoryMetadataProvider, which is a simple provider with no durability
intended for testing. Both the common and FS tests use this provider
rather than the FS provider. MemoryMetadataProvider is tested by
TestMemoryMetadataProvider, which inherits from the new common set,
TestMetadataProviders.

The common tests include checks that verify IllegalArgumentException is
thrown, which has been added to the API documentation in the interface.
Similar additions haven't been made to the FS MetadataProvider yet.