CDK-139: Removing Configuration from descriptors.
This commit removes the optional Configuration from the
DatasetDescriptor interface. Instead, optional properties can be set on
the descriptor. To avoid assuming the FileSystem for a Dataset, the
metadata for FileSystem-based data sets now includes a property,
cdk.filesystem.uri, which is used to open the correct FileSystem for
each DatasetDescriptor.

Removing the Configuration from the descriptor interface requires adding
it back to FileSystemDatasetRepository and FileSystemDataset.Builder, so
that those classes can open the correct FileSystem from the descriptor's
FS URI.

This also exposed a bug in Parquet, where new Configuration() is called
in a constructor, which causes parquet to get the wrong file system and
fail tests. There isn't a constructor that takes a FS or a
Configuration, so I'm not sure how to fix this. It doesn't happen when
the default FS from the environment is used.