CDK-159: Adding URI registration.
This adds DatasetRepositories#register, which is called to register a
URIPattern with an OptionBuilder that can build DatasetRepository
instances for the URI pattern's options. Now, DatasetRepositories no
longer depends on any particular implementation and the Hive/HCatalog
implementation can be supported.

DatasetRepository instances are loaded via ServiceLoader (like
FileSystem implementations) and register one or more URIs by calling
DatasetRepositories#register in a static block. Implementations should
be listed in META-INF/services/com.cloudera.cdk.data.DatasetRepository
files in their containing jars.

The new OptionBuilder interface defines a method for building an
instance from a Map of options rather than through a series of method
calls. This is necessary because options from URIPatterns are returned
all as a single object, the result of the match.

TestDatasetRepositories is now filesystem/TestFileSystemURIs because it
is specific to the FileSystem implementation.