CDK-139: Removing cdk.filesystem.uri property.
The information in the cdk.filesystem.uri property can be encoded in the
Dataset's locaton URI, and should be. To remove the URI for the file
system, this adds checks when creating DatasetDescriptors that some FS
information is in the location URI (it has a scheme). Most of the
changes in this commit are updating tests for this change.

Also, this moves the TestFileSystemRepoWithProvider test into the
filesystem package where it should be.