CDK-139: Adding FS back to the FS repo builder.
This will cause an error if the root and fs don't match, but I'm not
sure we want to deprecate it quite yet.

Right now, the FS repo relies on the FS metadata provider to set paths,
but the provider can accept paths passed in on the descriptor. It isn't
currently clear which one should set the final path for a dataset. The
repository is the natural place if there is good separation between the
two, but the Hive managed implementation must get the path from Hive. I
think the right thing to do for now is to leave this in so we can change
it later if we need to.