Adding Hive storage URIs.
New hive storage URIs:
* hive => a managed hive repository
* hive:/tmp/repo-path => an external hive repo with the given path
* hive:/tmp/repo-path?hdfs-host=<h>&hdfs-port=<p> => above + HDFS info

The Hive metastore connection *must* be set up in the environment
because there is no way to pass the options via URI and the URI system
does not pass Configuration instances. This is because we want to have
Hive URIs working but don't know what they should look like yet. The
authority section of the hive:<path> URIs is reserved for this
connection information.

To make the scheme-less URI work, this adds minor changes to the
URIPattern class and tests to allow and check null schemes.

Hive URI registration works exactly like FS registration, with a
impl.Loader class that contains all of the needed OptionBuilders. The
Loader is listed in the cdk-data-hcatalog jar's services.