Updating FS URIs to fix minor problems.
When the URI path is relative and empty or null, it should be resolved
to the current directory, not the FS root.

Because URIs are matched in order of registration, the opaque URIs
should be added last so they don't accidentally consume valid
hierarchical URIs.

Also, this fixes the use of FileSystemDatasetRepository.Builder, which
should not use fileSystem. It now uses qualified paths instead.