CDK-175: Adding support for Parquet with Hive, Impala.
This adds utiltiites to convert from Avro schemas to Hive column
descriptions, FieldSchema and TypeInfo. This also adds serde and
inputformat/outputformat class metadata to the convertion maps for
Parquet. Both Hive and Impala work with a simple record test schema.

For testing, it was necessary to exclude the versions of hive and hadoop
parquet-hive from parqet-hive's dependencies. The hcatalog module no
longer depends on parquet-hive, but adding the exclusion will prevent
needing to figure it out later.