CDK-177: Removing additions to PartitionStrategy.
Earlier commits for CDK-177 added keyFor methods to replace
partitionKeyForEntity, but Keys are going to be in the SPI rather than
the API. This means that the Key class defines these methods (through a
Builder) rather than exposing them from PartitionStrategy. This commit
removes the new keyFor(Object) methods and marks newKey as private. It
also removes references to keyFor(Object). keyFor(Marker) remains
because there is no equivalent method to get a PartitionKey; this should
be removed when PartitionKey is replaced.