CDK-177: Add FS partition enumerating util classes.
MultiLevelIterator is an abstract iterator that produces lists of a
given depth in a nested hierarchy, like a FileSystem. The
FileSystemPartitionIterator uses it to enumerate all paths in the
directory structure produced by a partition strategy.

The FS iterator also limits the paths it yields to those within a given
MarkerRange and uses the PathConversion class to convert between Path
and Key.  PathConversion is a replacement for the toDirectoryName and
fromDirectoryName methods that does not rely on FieldPartitioners to
convert between values and strings (instead it uses Conversions).

Also adding a null check to Key.