CDK-199: Deprecate valueToString, valueFromString.
This commit deprecates the valueToString and valueFromString methods on
FieldPartitioner. These methods are replaced in the implementation by
filesystem-specific methods in filesystem.PathConversion. The purpose of
this change is to avoid serialization dictated by the FieldPartitioners
and allow the Dataset implementations to handle it.

This removes the limitation that identity partiioners can work with only
String, Integer, or Long types. PathConversion uses Conversions and now
supports float, double, int, long, and string types.

TestCalendarFieldPartitioner has been replaced by TestPathConversion.