CDK-177: Update FS writers for Views.
This commit updates the FS writers to stop using the Dataset partition
functions in favor of views. The major change is that there is no longer
a hierarchy of PartitionedDatasetWriters, but one per View (if
partitioned).

Before this, PartitionedDatasetWriter would use the partition dataset to
instantiate inner writer. Now, the backing View does not change and it
instantiates file writers directly rather than relying on the inner
Dataset to not be partitioned (at some level) and finally creating the
file writers. A consequence is that data file writers need to be
instantiated by both the View (if not partitioned) and by the
PartitionedDatasetWriter (otherwise). This commit introduces
FileSystemWriters#newFileWriter to instantiate file-based writers for
both.

One last change: the deprecated constructor for PartitionedDatasetWriter
is has been removed rather than updated before being removed later.

This also re-enables the testLimitedWriter test, which is now passing.