This is a combination of 2 commits.
Added a new RandomAccessDatasetRepository interface that is a sub interface of
DatasetRepository, but overrides load, create, and update to return the
RandomAccessDataset type instead.

Created two new static methods on DatasetRepositories called openRandomAcces,
which return a RandomAccessDatasetRepsotory.

Added more thorough tests for the different range combinations in the
DaoViewTest test class.