CDK-177: Deprecate Dataset's partition methods.
With the addition of View#deleteAll, the views interface can handle all
of the tasks supported by the partition-specific methods. This commit
deprecates the partition-focused methods, which will be removed in
0.10.0.

This commit also includes updates to classes that use the deprecated
methods, like TestFileSystemDataset, and either adds warning
supression or fixes the implementation.