CDK-215: Use URI authority to connect to MetaStore.
This adds two new URI patterns to the Hive support that override the
hive.metastore.uris property using info from the URI authority section.
The URI authority is defaulted to the original value of
hive.metastore.uris, and only set in the Configuration if the host is
set in the URI match arguments.

The largest change in this commit is that URIs with the / as the path
section will produce managed Hive repositories. This is to enable
setting the host and port in the URI because URIs with authority set
must be absolute. Any path other than / will still result in an external
repository.

This also makes fileSystemURI methods static and uses an
UNSPECIFIED_PORT constant rather than -1.