CDK-245: Delete of a record from an HBase Dataset will delete the entire system column family, which will destroy needed metadata if more than one dataset is defined for the table
The ManagedSchemaEntityVersion avro schema no longer uses a keyAsColumn mapping to map
the schema version to the system column family. Instead, each type of schema for a table
gets its own custom version of this ManagedSchemaEntityVersion schema which has a
mapping type of Column, mapping it to a column qualifier with the schema name in it.

Added a composite test to the ManagedDaoTest to validate that deletes of composite
subentities no longer deletes the entire system column family.