Update data-shell.
In the scala file, package names are updated.

In the shell script, this commit uses /bin/bash instead of /bin/sh
because type -p is not defined for all shells. Also, the path to
data-shell-inc.scala is updated because the shell script expects to be
run from the root of one of the submodules.