CDK-317: Fix HCatalog URIs.
This updates URIPattern to catch the case where a URI pattern does not
have a path but the tested URI does, and fail the URI. This allows the
HCatalog URIs to separately handle URIs without a path (Managed) and
URIs with a path (External) while still allowing the user to override
the authority information, host/port.

Previously, the HCatalog URIs used the root path, /, to signal that
the URI was for a managed HCatalog repository rather than an external
one. This is no longer necessary because the update to URIPattern and
any non-null path now returns an external repo.