Add binary compatibility check in maven.
This adds the org.semver rule for the maven-enforcer-plugin to the
build. Because the plug-in doesn't appear to reliably honor the include
and exclude configuration (works for filesystem.* but not others) this
doesn't apply the enforcer rule automatically. Instead, it can be
activated for any build that uses the maven-enforcer-plugin by running
with: -Dcheck.compat=true.