CDK-354: Add View#deleteAll and FS impl.
This adds View#deleteAll to the public API. AbstractRefinableView
defaults the behavior to reject all deletes. The FS implementation
checks that a view's constraints are aligned with boundaries in the
partition strategy. This guarantees that there are no entities in the
view's covering partitions that do not match the view's constraints.