CDK-358: Add parquet-hive dependency to Hive implementation.
When loading a table, Hive tries to load the InputFormat class, so it
must be present for all client libraries. While it isn't needed for
tests to pass, without it kite-data-crunch tests fail. This avoids the
issue for all downstream clients.

This also excludes the hive-serde dependency from parquet-hive. This is
already a provided dependency for kite-data-hcatalog, so there is no
need to inherit it.