CDK-371: Update delete docs and consistency.
This updates the documentation for DatasetRepository#delete to be
explicit that both data and metadata will be removed. It also clarifies
the ambiguity between returning false and throwing
DatasetNotFoundException. Returning false is correct (enforced by
testing) so the docs now reflect this and HBase has been updated to only
return false rather than throwing an exception.