CDK-292: Throw SchemaNotFoundException instead of NPE.
This fixes the case where the schema being written is not in the schema
manager because the Dataset has not yet been updated. The NPE was thrown
when calling mapFromEntity on a null entity mapper because the schema
version (used to get a mapper) was -1.

Now, when getting a specific version number from a schema,
SchemaNotFoundException is thrown instead of returning -1 to signal that
the version was not present. This is consistent with the existing
documentation for SchemaManager#getEntityVersion.

This update also adds SchemaManager#hasSchemaVersion to check if
getEntityVersion would throw SchemaNotFoundException. This is used in
place of getEntityVersion where -1 was used to check if a version was
known to the schema manager.