CDK-320: Add MetadataProvider interface for compatibility.
This adds the org.kitesdk.data.MetadataProvider interface back, which
extends the SPI MetadataProvider for compatibility until 0.14.0. This
also deprecates MetadataProviderException. In 0.14.0, it can also be
removed.

Deprecation warnings for MetadataProviderException have been fixed by
catching DatasetException or a more appropriate exception and by
wrapping a more specific exception in a MetadataProviderException when
propagating exceptions. In most cases, a MetadataProviderException now
wraps a DatasetIOException. Wrapping exceptions is done by
Accessor#providerExceptionFor to avoid deprecation errors for imports,
which can't be suppressed.