CDK-320: Remove MetadataProvider from FS repo builder.
This removes the metadataProvider method that configures the FS
repository builder because the provider is no longer in the public API.
Callers setting the provider now must use the constructor. Because the
Builder can't be used to create a FS repository with a different
provider, the Hive integration now subclasses the FS repository and sets
the provider in its constructor.

FileSystemDatasetRepository#getMetadataProvider is now package-private
because it is required by the Hive intregration, which includes its own
package-private version.