Add runtime classpath building in executable.
This adds an initialization script based on sqoop's configuration script
in the header of the executable jar. HBase, HCatalog/Hive, and Zookeeper
classpaths are initialized.

This also changes the way the executable jar is built. The maven
plugin's option to include a custom script, scriptFile, wasn't being set
by maven, so this commit replaces it with a small script to build the
executable jar. This script is run with the maven-exec-plugin.