Minimize dataset executable jar.
This configures the maven-shade-plugin to minimize the size of the final
executable jar. Dependencies that are provided on the CDH 4.4 VM are
explicitly marked as provided and bundled dependencies (such as crunch
and findbugs annotations) are explicitly marked as compile.

Unfortunately, the maven-shade-plugin is including some transitive
dependencies, like hadoop-client, despite those dependencies being
included by a provided dependency, like kite-hadoop-dependencies. I
haven't found a work-around other than to specifically include a minimum
set of artifacts.