Fix avro conflict for hadoop1 and hadoop2.
This moves the avro dependency _above_ the hive dependencies in the
kite-data-hcatalog pom because hive-exec includes avro classes (Schema)
and is the cause of a java.lang.VerifyError. Moving the dependency
causes the constructed classpath to search the avro jar first.

This also marks hive-exec and hive-serde as provided dependencies,
although they are not. They are used directly, but their inclusion
breaks modules that depend on hcatalog because avro is included. This
means that other modules (like kite-data-crunch) have to include them
explicitly to run some tests. kite-tools already specifically marks
these as provided.

Last, this excludes avro and avro-mapred from hive-serde because maven
is including the 1.7.1 version in the classpath, even if the dependency
tree reports it as version managed.