CDK-94: Add kite-data-all convenience bundle.
This is *not* part of the build because there are packaging problems.

First, hive-exec is provided in kite-data-hcatalog because it shades
core avro classes that conflict and cause very bad runtime Errors.

Second, this may not be very useful because it doesn't help eliminate
transitive dependencies. This could mark packages provided, as in the
kite-tools shaded jar. The problem is that as an external dependency,
anything marked provided by this pom will not be included as a
transitive dependency and downstream projects either have to include
these specifically to compile (not helpful) or may introduce runtime
bugs because libraries are missing (worse). This may be more appropriate
as a parent pom.