CDK-367: Fix Conversions for Object.
Conversions was testing the expected type using isAssignableFrom, so the
first branch was always being taken because everything can be assigned
to Object. The fix is to test for Object specifically and return the
input object because no conversion is requested. This was exposed by
Marker's getAs function that is using an id partitioner's getType rather
than getType(inType).