CDK-367: Add methods to remove schema-embedded config.
This adds removeEmbeddedMapping and removeEmbeddedStrategy that remove
the mapping and strategy embedded in schemas. This is needed to clean up
schemas when they are used elsewhere, like in the CSV import code where
the destination dataset schema is used as the CSV read schema. Setting
the schema on the temporary CSV dataset should not also set a partiton
strategy and column mapping.

Also, this consolidates JSON parsing and error handling in JsonUtil.