CDK-385: Update URIPattern for Dataset URIs.
This updates URIPattern so that it can match after glob patterns. For
example, file:/*path/:dataset-name, will now match "dataset-name" before
matching the path glob. This makes it possible to define URI patterns
that include a globbed path component that ends in a dataset name.

This also adds sample use, which needs to be updated.