CDK-469: Move HDFS URL handler to SPI, make Configurable.
The URL handler was calling new Configuration() to get a config. This
was making tests fail when a schema URI is passed to DatasetDescriptor
and the HDFS information is not configured in the environment. The
solution is to allow the CLI to set the correct Configuration that
instances will use.