CDK-453: Move URI construction into FS repositories.
Some tests were failing because the repository did not have a URI and
could not use it to construct dataset URIs. This moves the repo URI
building into the repository's constructor rather than just using the
builder. This ensures repositories will always have a URI that can be
parsed by URIPattern.