CDK-484: Add Datasets URI tests and Datasets#exists.
This also fixes inconsistencies in how view URIs are handled. Create or
load operations will return View objects for view URIs and create will
create the underlying dataset correctly. Delete, update, and exists will
reject view URIs with an IllegalArgumentException. Update applies to an
entire dataset, so it doesn't make sense to allow it to be applied to a
view. Delete and exists could be applied to views in the future, but
this is not implemented (and may not be)