Fix exception swallowing in CLI commands.
The CLI commands were using a try/catch block to determine if the
incoming dataset identifier was a URI or a name. If the URI was a
dataset URI but another operation failed with IllegalArgumentException,
for example a problem during creation, then the catch would retry the
operation assuming the input was a dataset name.