Add mapping-config command to create column mappings.
This works like the partition-config command, where arguments are:
* <source>:key for key mappings
* <source>:version for occVersion mappings
* <source>:<col-family>[:<col-qual>] for all other mappings

The mapping type is determined by the schema type for non-key and
non-OCC mappings. If the field is an int or a long, then a counter is
used. If it is a record or a map, key-as-column is used. Otherwise, the
field is column-mapped.

If the column qualifier is missing for counter or column mappings, then
the source name is used.