Hide dataset repository selection options.
Before dataset URIs were supported, the CLI selected the dataset
repository based on options like --use-hbase and --directory. This was
limiting because only one repository could be selected. Now that URIs
can be used to identify a dataset in a single option, this is no longer
necessary. The old behavior is still around for backward-compatibility
and for ease of use (Hive is the default), but there is no need to keep
these options in the help menus.