Add DurableParquetAppender.
This appender wraps an AvroAppender and a ParquetAppender to provide
durability by writing to Avro in tandem with writing parquet files
directly. After the file is closed, the avro copy is removed in a new
cleanup phase in the appender. Any failure requires recovery using the
Avro files.

The default is to use the regular ParquetAppender and the durable
appender can be enabled by setting the descriptor property
"kite.parquet.durable-writes" to true.