CDK-488: Update to 2.3.0 to make copy test pass.
This updates to hadoop 2.3.0 so that the LocalJobRunner uses more than
one reducer and passes when we check that the correct number of files
are created. This uses the LocalJobRunner.setLocalMaxRunningReduces, and
does not run the test if it is not present so that the test is ignored
in the cdh4 and hadoop-1 profiles.